add UserOverrides to prototype metadata
